+++ CHANGES.rst
@@ -1,6 +1,7 @@
 1.0.0 (Unreleased)
 ==================
 
+- Add "ignore tags" option, which allows you to filter out various tags
 - Use AniBridge mappings to mop up missed Sonarr/Radarr titles
 
 0.9.0 (2025-09-13)